py2to3 compatibility fixes
Former-commit-id: 493d64d3a9d5ded4515067dcd05a0012e9483b7d [formerly 8f239626c8526b0b8ab6afad1400fbcdf49bc4f7]
Former-commit-id: 0934b9fa4560aee979d711f86aca82bb9330b37a

+++ tomopy/misc/morph.py
@@ -187,9 +187,9 @@ def _sample(arr, level, axis, mode):
     dx, dy, dz = arr.shape
 
     if mode == 0:
-        dim = arr.shape[axis] / np.power(2, level)
+        dim = int(arr.shape[axis] / np.power(2, level))
     if mode == 1:
-        dim = arr.shape[axis] * np.power(2, level)
+        dim = int(arr.shape[axis] * np.power(2, level))
 
     out = _init_out(arr, axis, dim)
     return extern.c_sample(mode, arr, dx, dy, dz, level, axis, out)
